Provide "TCE" for IVars and Futures/Promises
We (1) introducing a per-thread schedule for unwinding, and (2)
merge IVars to avoid space leaks in "tail recursive" scenarios
(flatMap).

By providing a schedule, the stack is unwound for each IVar.get.
The implementation of IVar.apply() (blocking get) is changed to get
around the implied deadlock situation.

IVar merging collapses semantically identical IVars into one. This
prevents space leaks in "tail-recursive" flatMaps:

  def loop() = ... flatMap { ... loop() }

here, in the old implementation, every iteration depends on the
next one, and thus we get a chain of linked Futures as long as the
number of iterations. This leads to an obvious space leak. By merging
early, we can collapse this chain, freeing references to the
intermediate Futures.

(sapling split of 0da72730a602435da71f77acacaab83d05a9d907)