finagle-memcached: Fix for CACHE-16 (requests with empty keys)
The symptom for CACHE-16 was that requests from finagle-memcached were
sometimes being made with empty keys. It turns out that this happens
when a request is retried. Because of an API wart in Netty, the tokens
that comprise a memcache command were being *consumed* rather than
*copied* when they were written to the output buffer. This means that
subsequent retries would find 0 bytes in the consumed token, thus the
empty keys.

This commit adjusts the encoder to call the method that copies the
bytes to the output buffer without changing the source buffer rather
than the one that consumes the source buffer.

For details on the difference between the two writeBytes methods, see:

<http://netty.io/docs/stable/api/org/jboss/netty/buffer/ChannelBuffer.html#writeBytes(org.jboss.netty.buffer.ChannelBuffer,%20int,%20int)>

and:

<http://netty.io/docs/unstable/api/org/jboss/netty/buffer/ChannelBuffer.html#writeBytes(org.jboss.netty.buffer.ChannelBuffer)>

(sapling split of 034347333f7d7cb292817b6ab5edcc812c87bcab)