tweetypie: Make the AdminHttpService accessible when running in Mesos
When a service registers itself in a ServerSet in Mesos, if it has an
endpoint labeled "http", the Mesos service proxy makes that HTTP
server accessible via a well-known name. This commit registers
TweetyPie's AdminHttpService as an additional "http" endpoint when it
registers with a ServerSet.

(sapling split of 8cd49c70259ae34ac84e84e74d508efd66cf82b6)