finagle-redis: return pairs for hgetall, zrangebyscores, zrevrangebyscores cmds
Change hGetAll, zRangeByScoresWithScores, and zRevRangeByScoresWithScores interfaces to return a map of the result pairs, as opposed to a sequence. This better reflects what the commands actually return, and the consumer doesn't have to extract the pairs on their end.

(sapling split of 8b8406a0e9c1ecfecf30fdf88c1539a45907b028)