util-core: Add efficient base 64 for longs
This specialized implementation is for use in cache keys, where we
want an ASCII representation, but we also want space-efficiency. This
implementation is not compatible with converting to a byte array and
then converting directly to base 64. It uses the base 64 alphabet, but
it represents the number as it would be represented in any other base,
from most to least significant digit, with leading zeros stripped.

Resulting maximum output String lengths:

  base 64: 11
  base 16: 16
  base 10: 20

Micro-benchmarked, this implementation is about as CPU efficient
and generates less garbage than converting to base 10 or base 16.

  base 64 (Base64Long.toBase64): 73.nanoseconds, 63.9 bytes of garbage
  base 16 (Long.toHexString)   : 67.nanoseconds, 208.0 bytes of garbage
  base 10 (Long.toString)      : 66.nanoseconds, 95.6 bytes of garbage

These measurements came from the Benchmark module from Decider's
tests.

RB_ID=66405

(sapling split of cc6b52f9957ecf5423226c4ddde055c896bed2dc)