Add Future.None
This is a helper for a common pattern in application code where you know that the
future need not be executed and its ultimate optional value will be missing.
Once this is in the main library, we can remove many places where this pattern is
already in use.

RB_ID=67072

(sapling split of c19a00c8e9716b6d614128481f709a54d352a4cd)