Changed Cache's internal implementation from queue to deque (to allow LIFO behavior)
Previous Cache implementation was a queue, which had FIFO behavior
with its entries. If we instead have LIFO behavior and the typical
working set of items is smaller than the size of the cache, then we
will likely use fewer cached items overall (O(working set) rather than
O(size of cache)).

Internally, the change was from a queue (FIFO behavior) to a deque
(LIFO behavior if configured as a stack, but with the ability to pull
items from the bottom of the stack as well, which was necessary when
expiring items).

Also added a test that checked that if the TTL has not been exceeded,
no items will be expired, and changed existing tests to LIFO
behavior.

RB_ID=78469

(sapling split of b4041c3f1572d307f2308b9409e20821c4a9f793)