Added name to HeapBalancer init (via exception) if HB size is 0
An empty HeapBalancer, when apply()ed, was returning a
NoBrokersAvailableException with no info (so it was difficult for a
user to tell which HeapBalancer failed).

Now HeapBalancer takes an exception arg during initialization. The
exception should be initialized with a string indicating the name of
the service, and when the exception is thrown for a HeapBalancer
that has no size, it shows the inited name. This is the same strategy
as TimeoutFilter.

RB_ID=88806

(sapling split of 88432136290cf1401469de0752044065c15be43e)