com.twitter.app: composable apps, flags.
This implements a very simple composable app mixins that allow
orthogonal components to mixed in as modules (one for logging is
included). It also specifies a flag implementation that is also very
simple but powerful and composable. Typeclasses are used for parsing
flags in a composable manner.

RB_ID=95957

(sapling split of e37ceccbdd1172a19bed64663f0e97e28b34b0da)