finagle: refactor clients and servers to be simpler, more logical.
Clients now comprise:

A Transporter[In, Out] is a function SocketAddress =>
Future[Transport[In, Out]]. That is, it connects to a socket address
and provides a typed transport to talk to that endpoint.

A Dispatcher[In, Out, Req, Rep] is a function Transport[In, Out] =>
Service[Req, Rep] that takes a transport and provides a service - a
bridge. This is what connects the "stream" oriented Transport to the
"RPC" oriented Service.

A Binder is a function SocketAddress => ServiceFactory, which is
pretty self-explanatory. Given a SocketAddress, bind it to a
ServiceFactory. From types, it's obvious that the role of the binder
is to establish the host-specific stack. In order to construct a
binder, you need a Transporter and a dispatcher:

	object MuxBinder extends DefaultBinder[ChannelBuffer, ChannelBuffer, ChannelBuffer, ChannelBuffer](
	  new Netty3Transporter(mux.PipelineFactory),
	  new mux.ClientDispatcher(_)
	)

And finally, the Client provides a Client[Req, Rep] given a Binder
(and possibly other configuration paramters, like a pool
implementation):

	object MuxClient extends DefaultClient[ChannelBuffer, ChannelBuffer](MuxBinder, _ => new ReusingPool(_))

Servers are simpler, with just two components:

A Listener[In, Out]:

	trait Listener[In, Out] {
	  def listen(addr: SocketAddress, newConnection: Transport[In, Out] => Unit): ListeningServer
	}

which is given a binding address and a function to call for each new
connection (with a a typed transport). Netty3Listeners are one such
example.

And the Server[In, Out, Req, Rep] which takes a listener and a
dispatcher, providing a Server[Req, Rep]:

	object MuxServer extends DefaultServer[ChannelBuffer, ChannelBuffer, ChannelBuffer, ChannelBuffer](
	  MuxListener, new mux.ServerDispatcher(_, _))

Taken all together, Mux.scala is now just [1]

*

I did a bunch of other refactors in order to support these changes.
Maybe some went a little too far. I had to stop myself in order to
limit the size of this change; so watch this space for more.

- finagle threadpools are now daemonized. To preserve the old behavior
for {Client,Server}Builder *only*, something called "ExitGuard" is
introduced, which does what it is named.

- Service, ServiceFactory, and a bunch of other things are made
Closable. The details of Closable are still a bit in flux. In
particular, it makes sense for Closables also to be Awaitable.

[1] Mux.scala:

	object MuxBinder extends DefaultBinder[ChannelBuffer, ChannelBuffer, ChannelBuffer, ChannelBuffer](
	  new Netty3Transporter(mux.PipelineFactory),
	  new mux.ClientDispatcher(_)
	)
	object MuxClient extends DefaultClient[ChannelBuffer, ChannelBuffer](MuxBinder, _ => new ReusingPool(_))

	object MuxListener extends Netty3Listener[ChannelBuffer, ChannelBuffer](mux.PipelineFactory)
	object MuxServer extends DefaultServer[ChannelBuffer, ChannelBuffer, ChannelBuffer, ChannelBuffer](
	  MuxListener, new mux.ServerDispatcher(_, _))

	/**
	 * A client and server for the mux protocol described in [[com.twitter.finagle.mux]].
	 */
	object Mux extends Client[ChannelBuffer, ChannelBuffer] with Server[ChannelBuffer, ChannelBuffer] {
	  def newClient(cluster: Cluster[SocketAddress]): ServiceFactory[ChannelBuffer, ChannelBuffer] =
	    MuxClient.newClient(cluster)

	  def serve(addr: SocketAddress, service: ServiceFactory[ChannelBuffer, ChannelBuffer]): ListeningServer =
	    MuxServer.serve(addr, service)
	}

RB_ID=117888
TBR=true

(sapling split of 7d9feec580f5ca1f3e7684c24e16c560b37b1d39)