cache the cancelled connections in WatermarkPool instead of CachingPool
CachingPool does not propagate interrupts of the service futures
and hopes to cache them when the connections are eventually
established. However, when no connections can be established in
a period of time a new connection will be created for every
new service call. The descried scenario can be exacerbated with
thrift where CachingPool only sees a connection attempt as
successful when the reply message of the protol up-negotiation
message is received. Moving the cancellation logic from
CachingPool to WatermarkPool fixes the described issue since
WatermarkPool does not allow # pending connections to exceed
the high watermark.

RB_ID=142028

(sapling split of c485646e187eab9186fd67533f412df451c7ea8a)