HeapBalancer: fix handling of unavailable nodes
Unavailable nodes were maintained in the heap, but skipped on
retrieval. However, the way in which they were skipped could cause
parts of the heap to be effectively shadowed, as nodes to the left of
unavailable nodes would not be scanned.

We change the way load accounting is done: instead of beginning at 0,
loads begin at Int.MinValue. Unavailable nodes are given a penalty of
-Int.MinValue, so unavailable nodes have loads > 0. We also maintain a
linked list of downed nodes. This list is scanned at the beginning of
every balance in order to see if they have been revived.

Together, these handle load balancing with failed nodes properly. In
the limit, if all nodes are failed, this will balance across all of
them evenly.

RB_ID=147920

(sapling split of eef836d770ace903621c696a3e2272b0923eea7a)