finagle-core: fix memory leak in Group#fromCluster
Group#fromCluster retained a reference to the head of
the spool. The spool was only used to propagate updates
which were subsequently committed to a separate collection.
This change moves the reference outside the class which
makes it available for the gc to reclaim.

RB_ID=179087

(sapling split of 3ea76692aae6d8e0e7f1e278fb52dbf552db7a70)