[geoduck] Use the new TransactionIngestionMetadata struct
We're now getting more information about the trustworthiness of the
client making the connection that causes the transaction (in
TransactionIngestionRequest.metadata). This change uses that data to
determine if the user's realtime position should be updated in the
cache.

RB_ID=222047

(sapling split of 75db79f447374373190749b67971ba8dbe9c4993)