[geoduck] filter results returned based on IP confidence rating
There are a few related changes here. In order to have the NetAcuity IP
confidence rating when we do reverse geocoding for clients it's necessary to
cache this data at ingestion time (as we don't necessarily have the IP at RGC
time). The code was designed in such a way that everything was using a
TransactionLocation as the domain object, which ties our internal representation
of the location to the object we're going to return to clients. This change a
TransactionLocation (the thrift object) from ServerTransactionLocation (a new
domain object), so that we can persist the confidence without exposing this
change to downstream users. The only time a TransactionLocation should come into
play is when we cross the outer perimeter and are returning a result to a
client.

The write flow looks like:

* A TransactionIngestionRequest is received from TFE
* The TransactionIngestionParser converts this into a ServerTransactionLocation
  * The ServerTransactionLocation contains a Confidence struct that holds the NetAcuity confidence values
* The ServerTransactionLocation is persisted, based on all previous rules
  * UserHandler does not take Confidence into account when deciding to update user location (open question)
  * All of the backend caches (User, Snowflake, Transaction) now take a ServerTransactionLocation and convert
    to TransactionLocationCacheEntry on the fly (which now contains a naConfidence field to hold these confidence
    values)

The read flow is essentially changed by adding a filter for PlaceTypes based on
confidence in the case that the source was LocationSource.Logical. Deciders can
be used to adjust the minimum value for accepting an IP coordinate as valid for
RGCing a country/region/city/zipcode. This filtering is used for all location
types (User, Snowflake, Transaction), and currently affects only IP sourced
entries.

RB_ID=228507

(sapling split of 6b9286c076f7653dba005d5c59233e2a5307658b)