util-core: added Promise.attached and Detachable This is the change formerly known as Promise.becomeUnlessInterrupted.
This will be used to restore us to the heady glory days of clean DelayedFactory, and will also be useful in Stu Hood's FutureCache.

If detach is called on a Promise[A] with Detachable, and then it is interrupted, then it will unlink the detached promise first, making it eligible for garbage collection.  Adding the detach() handle is very flexible, and means that you can use other flow control than interruption to remove it.

RB_ID=257735

(sapling split of ba289c4c78fa2f6e8c55cbbd1117d6c61df63f26)