[finagle] fix dependency problem with multiple version of serversets
Motivation:
After my last commit, (and with a empty ivy cache) the sbt doc compilation was broken in finagle-serversets.
I think (not sure) that's because of the version in the ivyXml was different than the one specified in the sbt.

Modification:
- Update all usage of c.t.common.zookeeper.server-set (finagle sbt & util sbt)

Result:
"./sbt +publish-local" is working!

RB_ID=295301

(sapling split of 40ede1c01e8902ecc7d3d28f14c10f9a22be4ee0)