Var: prevent stale updates
A previous change removed the check for stale observers. This adds it
back, together with a unittest.

We also ensure that closes are done synchronously in flatMap. Though
this is safe -- as long as you are not creating your own custom Var
that does not close synchronously -- we may consider converting to an
asynchronous propagation model to allow for this.

RB_ID=297347

(sapling split of 3e7245dc8e92852b5c8e5912ed64b831bd1eaf5a)