Event: mergeMap, not flatMap
Event's flatMap follows the definition of Var, but this is both
confusing and arguably wrong. This is further witnessed by the fact
that, were you to use flatMap with an aggregated Event, the aggregate
would be reset by each outer flatMap iteration.

Arguably, following the logic that an Event is a sequence-like
collection, flatMap should concatenate its inner Events. However,
Events are by definition unbounded. Thus I argue that there is no good
definiton of Event#flatMap.

However, we can define Event#mergeMap, which is the concurrent version
of the above -- instead of concatenating, we merge.

RB_ID=297875

(sapling split of 4025cbb09d07a550e2e42d796efacffced2f4606)