Finagle: Add Mux and ThriftMux Clients based on com.twitter.finagle.Stack
In the short term, in order to roll out Mux support in ClientBuilder we need to
provide Mux protocol `Stack`s. The crux of this change is the addition of
`MuxNetty3Stack` and `ReusingPoolModule` for use in creating a Stack-based
MuxClient. This is then used to provide a Stack-based ThriftMuxClient. In
support of this, I added a `Stack.Role` construct that takes the place of the
inconsistent `id` strings used in Stack/Stackable.

RB_ID=276105

(sapling split of 701ad06578a252c10a6880056135c10d3ff6ae0b)