finagle-serversets Stabilizer: don't consider Pending update successful
The SS2 stabilizer ticked limbo states on Pending: if a serverset was
pending for two TTLs, it would be ticked out. This is just wrong; this
change fixes this by separating Bound and Neg explicitly, and treating
everything else as a failure.

RB_ID=315439

(sapling split of 70940877de0ce08b4973c8fbcc13ff2c5b5598d9)