finagle: RetryPolicy filter, limit, combine I moved the RetryPolicy and Backoff classes out of the RetryFilter.scala file and into RetryPolicy.scala.  This makes it look like the whole thing is new, but the only additions are the instance methods .filter, .filterEach, and .limit, and the object method .combine.
The goal of filter, filterEach, and limit is to provide a means for assembling dynamic retry polices that can adapt retry strategies to back-pressure signals.

The goal of the combine method is to make it easier to build more complex retry policies from simpler policies; specifically, to make it easy to have different retry counts for different exception types (for example, 2 retries for WriteExceptions and 1 for TimeoutException), but with interleaved handling.  That way, it can handle a WriteException followed by a TimeoutException with the appropriate backoffs.

RB_ID=327379

(sapling split of ab7f9318e9b9e6f514da5b351442efcc63f9a48a)