Finagle: README refresh
Problem
Finagle's README is unreasonably long and outdated. The parts that are
still relevant and correct are duplicated by the user guide. The README
should be as simple as possible and delegate all actual documentation to
the user guide and API reference.

Solution
98% of the current README's content is deleted in favor of a summary of
Finagle, pointers to further documentation and support channels, and a
blurb about how to contribute to the project.
 deleted in favor of a summary of Finagle, pointers to further documentation and support channels, and a blurb about how to contribute to the project. I also replaced the "medium"-sized logo for the small one. IMO, the current image overly dominates the page.

RB_ID=329933

(sapling split of b7165d21650c29744ba195a8691a641528db6132)