Introduce 2-level caching in the name interpreter
Motivation

The current one-level caching scheme for interpretations of names is
too coarse grained. Since it is done on a Namer basis, any time there
is any Dtab change at all, a whole new set of clients may be created.
This causes latency issues when using Dtab overrides in servers with a
large number of clients.

Modifications

We change Namer to resolve down to Names instead of Paths that are
interpreted specially. We treat these names as a canonical
representation of a target. We separate binding and evaluation of
NameTrees.

We then introduce a 2-level caching scheme that maintains a client per
Name-set, and a client per Namer. This allows us to share clients at
the granularity of the Name, and to avoid NameTree evaluation at the
granularity of the Namer.

Result

Overrides are fast as they don't introduce additional client creations
due to the coarse nature of the current caching scheme.

The manner in which the change was done -- separating NameTree binding
from evaluation -- requires an eager treatment of NameTree binding.
This in turn means we need to fully explore all branches of Alt-nodes
before evaluating the resulting tree. In our current production
scenarios, that shouldn't matter, but may for future use cases. This
design decision should be reconsidered at this time.

RB_ID=325867
TBR=true

(sapling split of 72c53e7687dd51d5ebc1bef5dc0aa134afb24f4c)