Ostrich: Add System.exit to the end of abortabortabort Problem
`/abortabortabort` is meant to be the last resort that definitely kills a process once attempts to quit have failed. Unfortunately, non-daemon threads and other threads that have not been encapsulated in Ostrich's lifecycle framework can escape this and prevent the process from exiting. See [0] for the latest example of this issue, which results in a zombie process.

Solution

Adds a pluggable System.exit to the end of the MesosHandler`s abortabortabort.

Result

Calling `/abortabortabort` without a preceding `/quitquitquit` should cause the process to fully exit. Unfortunately, the science and ostrich implementations of `/quitquitquit` both kill threads that are necessary to serve a proceeding `/abortabortabort`, but we don't attempt to handle that here due to the deprecation of ostrich in favor of twitter-server, which ver has already improved.

[0] https://jira.twitter.biz/browse/RR-304

RB_ID=328023

(sapling split of cbaba285f41005f0cdb24391336c18089f719e83)