finagle: Add a new sbt target to publish locally to m2
Problem

Publishing the maven-scrooge-plugin requires having access to the latest finagle.
Every time we publish, we need to wait for the finagle release to be available on Sonatype.

Solution

Rather than waiting for the release to be available on Sonatype, this new sbt target
installs it locally.

Result

./sbt +publish-m2 publish the jars in the local maven repository

RB_ID=337149

(sapling split of cbd58d5a604cd9d20704444b6fd810dcb08cd4d8)