Use scala.Option in all com.twitter.scrooge files to avoid conflict with com.twitter.scrooge.Option
scrooge-generator Compiler refers to 'Option'. Normally, this refers
to 'scala.Option'. But if scrooge-runtime is loaded, it refers to
'com.twitter.scrooge.Option'. This can cause problems with projects
that include both scrooge-generator and scrooge-runtime, depending on
the order of jars in the classpath.

RB_ID=343137

(sapling split of b6dbc2d548d56e6e7944ff70583d5bcc69530844)