finagle-core: Invert the `cancelOnHangup` value passed to MaskCancelFilter.Param
Problem
In a recent change that removed `StackServer.transformed`, we inverted
the interpretation of the boolean Param that configures whether or not to
include a MaskCancelFilter. Since this inversion was not reflected in
`ServerBuilder.cancelOnHangup`, we effectively inverted the cancellation
behavior of any services that use ServerBuilder.

Solution
Invert the boolean argument passed to `MaskCancelFilter.Param` in
`ServerBuilder.cancelOnHangup`.

RB_ID=349129

(sapling split of 7f453b72246878dabb1175f9b039438fabc3ce22)