finagle-mux: refactor Session to make closing a Session uniform
Problem

MuxSession.connect and MuxSession.listen both return a (Session => Session). This makes closing a session and the releasing of resources difficult and non-obvious. You can close either instance of Session to terminate the Session, but cleaning up resources within the other instance of Session is difficult.

Solution

Rename Session to MuxService, and add a new Session which dispatches received messages and is provided to both connectors and listeners. If you wish to terminate a Session, you call close on Session. This will call close on the instances of MuxService, and then close the transport. Session.client (an instance of MuxService) is used to send rpcs/messages/pings/drain to the other party.

Result

The new interface provides a single way to close a session and clean up resources within an instance of MuxService.

RB_ID=317129

(sapling split of 6e995e99fc98de7f46587b8ae695ea17e371515e)