twitter-server: productivity = cpuTime/wallTime
Problem

The 'productivity' metric in twitter-server is pretty much useless. It
only measures the ratio of cpu time spent in user mode vs. system
mode, it does not account for blocking.

Solution

We can make a gross-ish estimation of this by measuring wall time. A
thread's cpuTime/wallTime measures the proportion of time the thread
was "doing something." This is by no means perfect: for example, the
kernel scheduler may simply choose not to run the (otherwise runnable)
thread, it doesn't account for preemption, etc. etc.

I think it is reasonable to define 'productivity' in those terms, with
the caveat that a lack of 'productivity' is not always attributable to
blocking alone.

The metric may be attenuated by a mismatch in accuracy between CPU
time (presumed getrusage) and walltime measurements (which can be done
by reading a timing register).

RB_ID=368879

(sapling split of 249a6d30597e570c61755bbfd52e52d7c5c98ba8)