Finagle: Improve Mux server close behavior, control messages to non-Mux clients
Problem

After `c.t.f.mux.Server#close` has been invoked, a Mux server will
respond to any Tdispatch messages with RdispatchNack. The
control-message handling code recently added to the ThriftMux
protocol-downgrading ChannelHandlers does not cover this message
type. This results in a race condition within the `mux.Server#close`
tests: if any of the test's un-Awaited outstanding requests are received
*after* `server.close()`, they'll trigger RdispatchNacks which will
cause the server's downgrading ChannelHandler to throw
UnexpectedRequestExceptions.

Additionally, `mux.Server#close` does not actually close the Transport
or Service, even once an Rdrain is received.

Solution

`mux.Server#close` will now close the underlying Transport if either a)
`closep` is satisfied (as in the case of receipt of an Rdrain), or b)
the argument deadline expires. Also a callback is added to
`trans.onClose` to close the Service.

The only means of negative acknowledgement afforded to ThriftMux servers
connected to non-Mux clients is to close the connection. Thus, the two
protocol-downgrading ChannelHandlers now close the channel when
RdispatchNack messages are issued. Furthermore, we remove the code that
simulated receipt of Rdrain in response to issued Tdrains.

RB_ID=360537

(sapling split of d332abbc37d8dd80253254cecc60c3ad256db96c)