finagle-mux: GC Avoidance Algorithm
Problem

In order to choose when we do GC avoidance, we need to guess in
advance when the JVM will do GC automatically, and respond properly.
As a first-order heuristic, we're going to use an exponential
weighted moving average.

Solution

We wrote a jvm GC heuristic for finagle, and use it to decide when
to drain.  After we've drained, we try to force a minor gc, and
then undrain.

Result

We have all of the pieces in place to detect how to set our leases
properly.

RB_ID=371522

(sapling split of 9d0a159307735211d62daad4b01bd8b6d76e9c00)