finagle-serversets: handle errors that occur when fetching endpoints
It appears that while ZK2 is pulling down endpoints for a service,
Zookeeper exceptions like NoNode in pulling data for any one of the
endpoints means resolution for the entire serverset goes negative.

Change the behavior so that if pulling the data for a single endpoint
raises a non-fatal exception, no data is returned for that endpoint.

RB_ID=378221

(sapling split of 83e6f766defc084b38d76266e11e4d15dffc18d5)