attach Dtab.local to NoBrokersAvailableException in filter Problem:
This change

  https://reviewboard.twitter.biz/r/418243

attempts to attach the local dtab to NoBrokersAvailableExceptions (for
display in error strings) when path names are used, but fails in some
cases because exceptions are created in a different thread-local context
from where they are thrown.

Solution:

Set the local dtab on NoBrokersAvailableExceptions in a post-filter
instead of at the point they're thrown.

Result:

Dtab is attached to exception in all cases.

NB: This should be replaced with a failure source once
NoBrokersAvailableException has been converted to a Failure.

RB_ID=426731
TBR=true

(sapling split of df88fc27eefc6653519cb8814f981e837275c465)