[finagle/wily] zipkin tracing for name resolution
Problem

We'd like to have zipkin data for wily name resolution. In particular,
we'd like to know the logical name requested for resolution, the
concrete name resolved, and the base and local dtabs.

Solution

Modify the DynNameFactory to take a tracing function, which will trace
either the bound name or the exception encountered during binding. I
don't super love putting it here, but it's the only place that has all
the necessary data.

I also had to move the Dtab.base ++ Dtab.local concatenation into the
dtabCache to make it easier to thread them through to the tracing
function.

I also considered having the DynNameFactory return both the Service and
the bound name; it was a little more clunky.

I'm not in love with the key names for the trace.

Result

Confirmed via println tracing that the traces are annotated as needed

RB_ID=426811

(sapling split of b51867f3b41809baebb721e19571defce58ac96b)