twitter-server: add gauge on eden allocations
Problem

It is difficult to estimate how many bytes an application is
allocating. With this information, it is easier to monitor for
regressions.

Solution

Added `Allocations` to TwitterServer which provides an `eden: Long`
method which returns an estimate of the total bytes allocated into the
eden. While not all allocations happen in the eden, it is the common
case.

If for some reason, the right diagnostic hooks are not available, -1L
is used as a sentinel value.

Result

Service owners have better visibility into their allocation rates.

RB_ID=429819

(sapling split of 404b796463278f2885ce1e84bff5eb3093a72219)