reusing pool: isAvailable === underlying.isAvailable
Service.isAvailable was overriden to check for service.isAvailable.
This is problematic because the load balancer will exclude traffic to
hosts where isAvailable==false; however, ReusingPool is happy to
reconnect when this is true. But with the current code (and load
balancer), this won't happen unless *all* connections are unavailable.

RB_ID=437701

(sapling split of ae0144e9d826a7d38e89a29074975acf7cff47eb)