util-cache: Adds a Guava-backed asynchronous cache
Problem

It's often useful to have an in memory, in process cache of futures.
However, it's a tricky to get right.  This means that the twitter
codebase is littered with slightly incorrect ones.  Even if you fix
one of them, because they're all different, other implementations
won't get the update.

Solution

Combining guava caches and a few simple primitives for caches, we
can make an easily understandable and easily modifiable cache for
futures.

Result

There's a canonical future cache that can be pointed to when people
need it

RB_ID=400407

(sapling split of 7d84355f0f8579c5420f56dca9dc2ca65724b3e5)