SingletonPool: a single-service pool with good semantics
Problem

The code for reusing pool was a bit difficult to disentangle; it
wrapped up the pool's state with a promise containing the current
service. Reasoning about the various edge cases was difficult.

Further, ReusingPool didn't define semantics for service close. The
idea behind this was that the underlying connection would close by
itself when a server goes away. However, this relies on the server
explicitly shutting down TCP connections (in the case of Mux), and it
seems like a good idea to have this case explicitly covered to better
account for resource usage.

Solution

We introduce SingletonPool which has roughly the same semantics as
ReusingPool, but with code that's much easier to reason about.
Further, we add proper closing semantics.

RB_ID=438275

(sapling split of 4d9bcc95c0825cf3c2fde172eed486cfbbfc67f1)