Add option for not building thrift from dependencies
This gives the user the option of extracting thrift files from
dependencies without building them, so they can be included in thrift
files in the current repo.

This is useful when the dependencies' thrift files were already built
with scrooge, when it would be redundant and a waste of time to rebuild
them upon extraction.

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

RB_ID=449089

(sapling split of 80de8dbeee8036a5b8fcbfadd791d34d254b476a)