Add support for non-URLClassloaders to LoadService (Fixes #304)
Problem

When the classloader is not a URLClassloader, it is not searched for
resources by ClassPath.  This causes discovery of services from
LoadService to render an incomplete list.

Solution

Merge the existing list of services found with ClassPath with a plain
old classloader.getResources().  Not the most pretty but it handles
cases where the classloader does not inherit from URLClassloader.

Result

Expected services are returned.

Signed-off-by: Steve Gury <stevegury@twitter.com>

RB_ID=435579

(sapling split of 8ee692d6c28b01297c8e27d734a2b3e655ca0575)