Fix Buf.ByteBuffer.slice
Problem

Buf.ByteBuffer.slice ignores position when updating limit.
Example: Allocate a ByteBuffer and construct a Buf from it.

    // pos=0 lim=3
    val b = Buf.ByteBuffer(java.nio.ByteBuffer.allocate(3))

Drop the first byte.

    // pos=1 lim=3
    b.slice(1, b.length)

Drop the first byte and get the first byte, this should be
equivalent to getting the second byte of the original b.

    // pos=1 lim=1
    b.slice(1, 3).slice(0, 1)

But because of the bug, the limit is updated to the wrong
index, resulting in an empty Buf.

Solution

Factor in the position when updating limit in a slice.

Result

Slice works again.

RB_ID=474895

(sapling split of 8a91a4dc7ac1050e85d9baa1eedd664ea2aeb940)