Trace, Context: add transitional APIs
In order to get the request context change[1] shipped, we need to
introduce a transitional API so that we can introduce it to science
without having an intermediate broken state.

I have a separate branch in science that makes changes according to
this, and they seem to be all that's required.

[1] https://reviewboard.twitter.biz/r/469081/

RB_ID=480925

(sapling split of 87358a65f6e56c4977b6dcbf986cc9dcf0401e12)